Rename folders to mitre-inspec-objects as well
