+++ lib/mitre-inspec-objects.rb
@@ -1,4 +1,4 @@
 # frozen_string_literal: true
 
-require "inspec-objects/version"
+require "mitre-inspec-objects/version"
 require_relative "inspec/objects"
