+++ lib/mitre-inspec-objects/version.rb
