+++ mitre-inspec-objects.gemspec
@@ -3,7 +3,7 @@
 lib = File.expand_path("lib", __dir__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
 
-require "inspec-objects/version"
+require "mitre-inspec-objects/version"
 
 Gem::Specification.new do |spec|
   spec.name          = "mitre-inspec-objects"
